minor bugfixes
