+++ client/src/css/monster.css
@@ -8,6 +8,7 @@
   grid-template-columns: 1fr 1fr 1fr;
   align-items: start;
   top: 20px;
+  padding-bottom: 60px;
 }
 
 .monster {
@@ -34,11 +35,20 @@
   font-size: 16px;
 }
 
-.form .category {
-  margin-bottom: 2px;
+.form {
+  display: flex;
+  flex-flow: row wrap;
+  justify-content: center;
+  align-items: flex-start;
+  align-content: flex-start;
 }
 
-.form .category:first-child {
+.form .category {
+  margin-bottom: 4px;
+  width: 100%;
+}
+
+.form .category:first-child, .form .category:nth-child(2) {
   margin-top: 15px;
 }
 
@@ -51,15 +61,13 @@
 }
 
 .form .category input:focus {
-  background-color: #a7e49d;
+  background-color: var(--mcl2);
 }
 
 .form button {
   position: relative;
   width: 25%;
-  left: 50%;
   bottom: -15px;
-  transform: translateX(-50%);
   margin: 10px 0;
   margin-top: 0;
   padding: 5px;
@@ -74,6 +82,23 @@
 .form button:hover {
   background-color: var(--mcl2);
   color: white;
+}
+
+.create_monster {
+  width: 100%;
+  height: 100vh;
+  position: fixed;
+  top: 0;
+  left: 0;
+  z-index: 1001;
+  margin: 0;
+  border-radius: 0;
+  border-color: transparent;
+}
+
+.create_monster .category {
+  margin: 4px 4px;
+  width: 45%;
 }
 
 .details_toolbar {
