+++ client/src/css/monster.scss
@@ -7,6 +7,7 @@
     grid-template-columns: 1fr 1fr 1fr;
     align-items: start;
     top: 20px;
+    padding-bottom: 60px;
   }
 }
 
@@ -37,12 +38,18 @@
     font-size: 16px;
   }
 }
-
 .form {
-  .category {
-    margin-bottom: 2px;
+  display: flex;
+  flex-flow: row wrap;
+  justify-content: center;
+  align-items: flex-start;
+  align-content: flex-start;
 
-    &:first-child {
+  .category {
+    margin-bottom: 4px;
+    width: 100%;
+
+    &:first-child, &:nth-child(2) {
       margin-top: 15px;
     }
 
@@ -55,16 +62,14 @@
     }
 
     input:focus {
-      background-color: #a7e49d;
+      background-color: var(--mcl2);
     }
   }
 
   button {
     position: relative;
     width: 25%;
-    left: 50%;
     bottom: -15px;
-    transform: translateX(-50%);
     margin: 10px 0;
     margin-top: 0;
     padding: 5px;
@@ -79,6 +84,23 @@
       background-color: var(--mcl2);
       color: white;
     }
+  }
+}
+
+.create_monster {
+  width: 100%;
+  height: 100vh;
+  position: fixed;
+  top:0;
+  left: 0;
+  z-index: 1001;
+  margin: 0;
+  border-radius: 0;
+  border-color: transparent;
+
+  .category {
+    margin: 4px 4px;
+    width: 45%;
   }
 }
